Clean Conv3DV2 Compilation Warning



# message auto-generated for no-merge-commit merge:
!336 merge conv3dv2-warning-cleanup into master

Clean Conv3DV2 Compilation Warning

Created-by: LiuZonggu
Commit-by: LiuZonggu
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
Compiler warnings occur due to type mismatches involving attrInfo_.groupOpt (declared as int64_t) in logging and comparison operations.

Logging: Updated the format specifier from %lu (unsigned long) to %ld (signed long) for printing attrInfo_.groupOpt to match its signed type.
Comparisons: Ensured type consistency by converting related variables in comparison operations to signed types (int64_t).

Eliminates compiler warnings by enforcing proper type alignment.
Maintains logical correctness in both logging and comparison behavior.

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->
Related to issue [#186](https://gitcode.com/cann/ops-nn/issues/186) 
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!336

+++ conv/conv3d_v2/op_host/op_tiling/conv3d_tiling_engine.cpp
@@ -732,7 +732,7 @@ bool Conv3dTilingEngine::CheckGroupOptAgainstWeightShape(uint64_t weightD, uint6
     if (expectedWeightD != weightD) {
         OP_LOGE(logTag_.c_str(),
                 "Weight D dimension mismatch: expected %lu, got %lu "
-                "(groupOpt=%lu, cinOpt=%lu, k0=%lu, kdkhkw=%lu)",
+                "(groupOpt=%ld, cinOpt=%lu, k0=%lu, kdkhkw=%lu)",
                 expectedWeightD, weightD, attrInfo_.groupOpt, shapeInfo_.cinOpt, k0, kdkhkw);
         return false;
     }
@@ -868,15 +868,16 @@ void Conv3dTilingEngine::CheckOutputShapeSizeLimits()
 void Conv3dTilingEngine::CheckAttrShapeSizeLimits()
 {
     // Check stride and dilation dimensions (legacy code checks these)
-    if (attrInfo_.strideD > MAX_ORI_ONE_DIM_SIZE) {
+    if (attrInfo_.strideD > static_cast<int64_t>(MAX_ORI_ONE_DIM_SIZE)) {
         OP_LOGW(logTag_.c_str(), "Conv3D AscendC: StrideD (%ld) is out of range[1, %lu].",
                 attrInfo_.strideD, MAX_ORI_ONE_DIM_SIZE);
     }
-    if (attrInfo_.dilationD > MAX_ORI_ONE_DIM_SIZE) {
+    if (attrInfo_.dilationD > static_cast<int64_t>(MAX_ORI_ONE_DIM_SIZE)) {
         OP_LOGW(logTag_.c_str(), "Conv3D AscendC: DilationD (%ld) is out of range[1, %lu].",
                 attrInfo_.dilationD, MAX_ORI_ONE_DIM_SIZE);
     }
-    if (attrInfo_.padHead > MAX_ORI_ONE_DIM_SIZE || attrInfo_.padTail > MAX_ORI_ONE_DIM_SIZE) {
+    if (attrInfo_.padHead > static_cast<int64_t>(MAX_ORI_ONE_DIM_SIZE) ||
+        attrInfo_.padTail > static_cast<int64_t>(MAX_ORI_ONE_DIM_SIZE)) {
         OP_LOGW(logTag_.c_str(), "Conv3D AscendC: PadHead/PadTail (%ld/%ld) is out of range[1, %lu].",
                 attrInfo_.padHead, attrInfo_.padTail, MAX_ORI_ONE_DIM_SIZE);
     }
